Nodes special icons and graph interconnection

+++ mobile/app/src/main/java/com/bellako/kiwi/features/nodes/model/INodesViewModel.kt
@@ -6,7 +6,7 @@ import com.bellako.kiwi.features.nodes.data.NodesState
 interface INodesViewModel : IBaseViewModel<NodesState> {
     fun loadNodes()
 
-    fun unlockNode(nodeId: Int)
+    fun unlockNode(nodeId: Long)
 
-    fun completeNode(nodeId: Int)
+    fun completeNode(nodeId: Long)
 }
